+++ app/services/spree_emerchantpay_genesis/notifications/service_handler.rb
@@ -1,5 +1,3 @@
-require 'spree/gateway/emerchantpay_direct'
-
 module SpreeEmerchantpayGenesis
   module Notifications
     # Notification Handler Service