+++ app/services/spree_emerchantpay_genesis/threeds/method_continue.rb
@@ -1,5 +1,3 @@
-require 'spree/gateway/emerchantpay_direct'
-
 module SpreeEmerchantpayGenesis
   module Threeds
     # Method Continue service