+++ spec/factories/emerchantpay_checkout_gateway_facotry.rb
@@ -7,15 +7,15 @@ FactoryBot.define do
     # be recorded on VCR, so they can be safely replaced with
     # placeholder afterwards
     transient do
-      username { Rails.application.secrets.username || 'example_username' }
-      password { Rails.application.secrets.password || 'example_password' }
+      genesis_credentials = Rails.application.credentials[:genesis]
+
+      username { genesis_credentials ? genesis_credentials[:username] : 'example_username' }
+      password { genesis_credentials ? genesis_credentials[:password] : 'example_password' }
       transaction_types { %w(authorize3d sale3d wechat post_finance trustly_sale) }
-      challenge_indicator { 'preference' }
-      test_mode { 'true' }
     end
 
     before(:create) do |gateway, evaluator|
-      %w(username password transaction_types challenge_indicator test_mode).each do |preference|
+      %w(username password transaction_types).each do |preference|
         gateway.__send__ "preferred_#{preference}=", evaluator.__send__(preference)
       end